Fix authorization for delete recipe

+++ pages/api/recipe/delete/[id].js
@@ -1,5 +1,5 @@
 import prisma from "../../../../lib/prisma.ts"
-import { checkAuth } from "../../../../lib/auth";
+import { checkAuth, checkIfCook } from "../../../../lib/auth";
 
 export default async (req, res) => {
   const isAuth = await checkAuth(req);
@@ -7,6 +7,14 @@ export default async (req, res) => {
     res.status(403).json({ err: "Forbidden" });
     return;
   }
+
+  const isTheOwner = await checkIfCook(req, req.body.ownerId);
+  if (!isTheOwner) {
+    res.status(403).json({ err: "Forbidden" });
+    return;
+  }
+
+
   const { id } = req.query;
   try {
     const deleteRecipe = await prisma.recipe.delete({
